chore: using now hr test helper
